feat: rewrite to ts controller.js
